Fix the CI for updated Github action virtual environments (#577)
* Try to fix some warnings.

* Fix out of bounds in turbinflow in PelePhysics and turn off some new clang-tidy checks.

* Update PelePhysics.

* Use concatenated namespace.

* Use older CMake for CUDA CI.

* Try to fix CUDA and HIP CI.

* Turn off more warnings.

* Update PelePhysics. Try macos-12 for cppcheck.

* Fix some narrowing conversions and turn off macro to enum warning.

* Fix implicit casts.

* Use .data() instead of the address of the first element of an array.

* Fix narrowing conversion.

+++ Source/Geometry.cpp
@@ -309,7 +309,8 @@ PolygonRevolution::build(
   amrex::Vector<amrex::EB2::PlaneIF> planes;
   for (int ipoly = 0; ipoly < num_poly; ipoly++) {
     const amrex::Vector<amrex::RealArray>& polygon = polygons[ipoly];
-    int numPts = polygon.size(); // Number of pts in this polygon
+    int numPts =
+      static_cast<int>(polygon.size()); // Number of pts in this polygon
     for (int n = 0; n < numPts; n++) {
       // The normal and point is space used to specify each half plane/space
       amrex::RealArray normal;
